Add delete multiple docs

+++ src/client/controller/SendMailController.java
@@ -12,7 +12,7 @@ import javafx.scene.layout.GridPane;
 import javafx.scene.text.Text;
 import javafx.stage.FileChooser;
 import client.main.ClientApp;
-import client.model.Mailsendreceivetest;
+import client.model.MailSendReceive;
 
 import java.io.File;
 import java.security.GeneralSecurityException;
@@ -50,22 +50,22 @@ public class SendMailController {
         FileChooser fileChooser = new FileChooser();
         fileChooser.setTitle("Select a file to upload");
         File file = fileChooser.showOpenDialog(clientApp.getPrimaryStage());
-        attachments.add(file);
+        if (file != null) {
+            attachments.add(file);
 
-        int row = attachments.size();
+            int row = attachments.size();
 
-        ImageView closeImg = new ImageView("File:src/images/closeBtn.png");
-        closeImg.setFitWidth(25.0);
-        closeImg.setFitHeight(22.0);
-        closeImg.setCursor(Cursor.HAND);
-        closeImg.setOnMouseClicked(e -> deleteFile(row-1));
+            ImageView closeImg = new ImageView("File:src/resources/img/closeBtn.png");
+            closeImg.setFitWidth(25.0);
+            closeImg.setFitHeight(22.0);
+            closeImg.setCursor(Cursor.HAND);
+            closeImg.setOnMouseClicked(e -> deleteFile(row-1));
 
-        ListView<GridPane> attachmentsList = new ListView<>();
-        GridPane grid = new GridPane();
-        grid.add(new Text(file.getName()), 0, 0);
-        grid.add(closeImg, 1, 0);
-        attachmentsGrid.add(grid, row, 0);
-
+            GridPane grid = new GridPane();
+            grid.add(new Text(file.getName()), 0, 0);
+            grid.add(closeImg, 1, 0);
+            attachmentsGrid.add(grid, row, 0);
+        }
     }
 
     @FXML
@@ -76,9 +76,9 @@ public class SendMailController {
     @FXML
     private void sendMail(ActionEvent event) {
         try {
-            if(attachments.isEmpty()) Mailsendreceivetest.sendmessage(this.clientApp.getUser(), this.toInput.getText(),
+            if(attachments.isEmpty()) MailSendReceive.sendMessage(this.clientApp.getUser(), this.toInput.getText(),
                     this.mailContent.getText(), this.subjectInput.getText());
-            else Mailsendreceivetest.sendmessagewithattachement(this.clientApp.getUser(), this.toInput.getText(),
+            else MailSendReceive.sendMessageWithAttachment(this.clientApp.getUser(), this.toInput.getText(),
                     this.mailContent.getText(), this.subjectInput.getText(), this.attachments, this.clientApp.getPp());
         } catch (GeneralSecurityException e) {
             e.printStackTrace();
@@ -86,7 +86,7 @@ public class SendMailController {
     }
 
     private void deleteFile(int id) {
-        ImageView closeImg = new ImageView("File:src/images/closeBtn.png");
+        ImageView closeImg = new ImageView("File:src/resources/img/closeBtn.png");
         closeImg.setFitWidth(25.0);
         closeImg.setFitHeight(22.0);
         closeImg.setCursor(Cursor.HAND);
